requirements.txt
requests>=2.31.0
pandas>=2.2.1
setuptools>=69.2.0
python-dotenv>=0.19.0